+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - PR #3776 Add column equivalence comparator (using epsilon for float equality)
 - PR #3667 Define and implement round-robin partition API.
 - PR #3690 Add bools_to_mask
+- PR #3761 Introduce a Frame class and make Index, DataFrame and Series subclasses
 - PR #3683 Added support for multiple delimiters in `nvtext.token_count()`
 - PR #3792 Adding is_nan and is_notnan